+++ freediscovery/text.py
@@ -11,13 +11,14 @@ import pandas as pd
 import scipy.sparse
 from sklearn.externals import joblib
 from sklearn.externals.joblib import Parallel, delayed
-from sklearn.feature_extraction.text import TfidfTransformer, TfidfVectorizer
+from sklearn.feature_extraction.text import CountVectorizer
 from sklearn.preprocessing import normalize
 from sklearn.pipeline import make_pipeline
 
 from ._version import __version__
 from .pipeline import PipelineFinder
 from .utils import generate_uuid, _rename_main_thread
+from .feature_weighting import FeatureWeightingTransformer, _validate_smart_notation
 from .ingestion import DocumentIndex
 from .preprocessing import processing_filters
 from .stop_words import _StopWordsWrapper
@@ -39,11 +40,9 @@ def _preprocess_stream(doc, steps=None):
     """ Apply pre-processing steps """
 
     if steps:
-        #doc = doc.splitlines()
         for key in steps:
             func = processing_filters[key]
-            doc = func(doc) #[func(line) for line in doc]
-        #doc = '\n'.join(doc)
+            doc = func(doc)
 
     return doc
 
@@ -60,12 +59,9 @@ def _vectorize_chunk(dsid_dir, k, pars, pretend=False):
 
     mslice = slice(k*chunk_size, min((k+1)*chunk_size, n_samples))
 
-    if pars['use_idf']:
-        pars['binary'] = False  # need to apply TFIDF weights first
-
     hash_opts = {key: vals for key, vals in pars.items()
                  if key in ['stop_words', 'n_features',
-                            'binary', 'analyser', 'ngram_range']}
+                            'analyser', 'ngram_range']}
     fe = HashingVectorizer(input='content', norm=None,
                            non_negative=True, **hash_opts)
     if pretend:
@@ -92,9 +88,9 @@ class FeatureVectorizer(object):
     """
 
     _PARS_SHORT = ['data_dir', 'n_samples', 'n_features',
-                   'n_jobs', 'chunk_size', 'norm',
+                   'n_jobs', 'chunk_size',
                    'analyzer', 'ngram_range', 'stop_words',
-                   'use_idf', 'sublinear_tf', 'binary', 'use_hashing',
+                   'weighting', 'use_hashing',
                    'creation_date']
 
     _wrapper_type = "vectorizer"
@@ -182,9 +178,8 @@ class FeatureVectorizer(object):
 
     def setup(self, n_features=None, chunk_size=5000, analyzer='word',
               ngram_range=(1, 1), stop_words=None, n_jobs=1,
-              use_idf=False, sublinear_tf=False,
-              binary=False, use_hashing=False,
-              norm='l2', min_df=0.0, max_df=1.0,
+              use_hashing=False,
+              weighting='nnc', min_df=0.0, max_df=1.0,
               parse_email_headers=False,
               preprocess=[]):
         """Initalize the features extraction.
@@ -226,14 +221,8 @@ class FeatureVectorizer(object):
         max_features : int or None, default=None or 100001
             If not None, build a vocabulary that only consider the top
             max_features ordered by term frequency across the corpus.
-        binary : boolean, default=False
-            If True, all non-zero term counts are set to 1. This does not mean
-            outputs will have only 0/1 values, only that the tf term in tf-idf
-            is binary. (Set idf and normalization to False to get 0/1 outputs.)
-        norm : 'l1', 'l2' or None, optional
-            Norm used to normalize term vectors. None for no normalization.
-        use_idf : boolean, default=True
-            Enable inverse-document-frequency reweighting.
+        weighting : str
+            SMART weighting type
         preprocess : list of strings, default: []
             A list of pre-processing steps, including 'emails_ingore_header'
         """
@@ -251,16 +240,17 @@ class FeatureVectorizer(object):
             raise WrongParameter('len(gram_range=={}!=2'
                                  .format(len(ngram_range)))
 
-        if norm != 'l2':
-            warnings.warn("the use of 'l2' norm is stronly advised;"
-                          "distance calculations"
-                          " may not be correct with other normalizations."
-                          " Currently norm={}".format(norm))
+        _, _, weighting_n = _validate_smart_notation(weighting)
+        if weighting_n == 'n':
+            warnings.warn('You should use either cosine or pivoted normalization '
+                          'i.e. weighting should be "**[cp]"',
+                          UserWarning)
+
         for key in preprocess:
             if key not in processing_filters:
                 raise WrongParameter(('Unknown preprocessing step {} '
                                       ' must of be of {}')
-                                      .format(key, ', '.join(list(processing_filters.keys()))))
+                                     .format(key, ', '.join(list(processing_filters.keys()))))
 
         if stop_words in [None, 'english', 'english_alphanumeric']:
             pass
@@ -285,10 +275,8 @@ class FeatureVectorizer(object):
                 'n_samples': None, "n_features": n_features,
                 'chunk_size': chunk_size, 'stop_words': stop_words,
                 'analyzer': analyzer, 'ngram_range': ngram_range,
-                'n_jobs': n_jobs, 'use_idf': use_idf,
-                'sublinear_tf': sublinear_tf,
-                'binary': binary, 'use_hashing': use_hashing,
-                'norm': norm, 'min_df': min_df, 'max_df': max_df,
+                'n_jobs': n_jobs, 'use_hashing': use_hashing,
+                'weighting': weighting, 'min_df': min_df, 'max_df': max_df,
                 'parse_email_headers': parse_email_headers,
                 'type': type(self).__name__,
                 'preprocess': preprocess,
@@ -489,27 +477,20 @@ class FeatureVectorizer(object):
 
                 res = self._aggregate_features()
 
-                if pars['use_idf']:
-                    tfidf = TfidfTransformer(norm=pars['norm'], use_idf=True,
-                                             sublinear_tf=pars['sublinear_tf'])
-                    res = tfidf.fit_transform(res)
-                    vect = make_pipeline(vect, tfidf)
                 self._vect = vect
             else:
                 opts_tfidf = {key: val for key, val in pars.items()
-                              if key in ['stop_words', 'use_idf',
+                              if key in ['stop_words',
                                          'ngram_range', 'analyzer',
-                                         'sublinear_tf',
                                          'min_df', 'max_df']}
 
-                tfidf = TfidfVectorizer(input='content',
-                                        max_features=pars['n_features'],
-                                        norm=pars['norm'],
-                                        **opts_tfidf)
+                vect = CountVectorizer(input='content',
+                                       max_features=pars['n_features'],
+                                       **opts_tfidf)
                 text_gen = (_preprocess_stream(_read_file(fname), pars['preprocess'])
                             for fname in pars['filenames_abs'])
-                res = tfidf.fit_transform(text_gen)
-                self._vect = tfidf
+                res = vect.fit_transform(text_gen)
+                self._vect = vect
             fname = dsid_dir / 'vectorizer'
             if self._pars['use_hashing']:
                 joblib.dump(self._vect, str(fname))
@@ -517,14 +498,9 @@ class FeatureVectorizer(object):
                 # faster for pure python objects
                 with fname.open('wb') as fh:
                     pickle.dump(self._vect, fh)
-
-            if pars['norm'] is not None:
-                res = normalize(res, norm=pars['norm'], copy=False)
-            else:
-                # scale feature to [0, 1]
-                # this is necessary e.g. for SVM
-                # and does not hurt anyway
-                res /= res.max()
+            fname = dsid_dir / 'weighting_transformer'
+            wt = FeatureWeightingTransformer(pars['weighting'])
+            res = wt.fit_transform(res)
 
             del self.pars_['filenames_abs']
 
@@ -664,14 +640,11 @@ class FeatureVectorizer(object):
     def n_features_(self):
         """ Number of features of the vecotorizer"""
         from sklearn.feature_extraction.text import HashingVectorizer
-        from sklearn.pipeline import Pipeline
         vect = self.vect_
         if hasattr(vect, 'vocabulary_'):
             return len(vect.vocabulary_)
         elif isinstance(vect, HashingVectorizer):
             return vect.get_params()['n_features']
-        elif isinstance(vect, Pipeline):
-            return vect.named_steps['hashingvectorizer'].get_params()['n_features']
         else:
             raise ValueError